* features/authenticate:
  GraphQL authorization.
  Allow path to facebook
  Protect accessing fb auth.
  Add authorise in fetching fbayth
  Login response admin object.
  Remove fixture when end of test
  Implement JWT Token decode and authorisation.
  Add test case for ensure express will protect resource when request without credential or wrong access level.
  allow login without auth
  Build before test
  Add auth testing.
  Append mocha exceed timeout.
  Add GraphQL schema.
  Add mutation registerOrLogin, input claim and profile, if login success, return jwtToken.
  Fix error in migration file and add column to user profile model.
  Add model into index.
  Define GraphQL type.
  Add migration to add column into user profile.

+++ src/data/mutations/deleteAdmin.js
@@ -8,7 +8,11 @@ const deleteAdmin = {
     id: { type: new GraphQLNonNull(GraphQLInt) },
   },
 
-  async resolve(obj, {id}) {
+  async resolve(obj, {id}, context, { rootValue: { request } }) {
+    if (!request.user) {
+      return false;
+    }
+
     const affectedRows = Admin.destroy({ where: { id: id } });
 
     return (affectedRows <= 0 ? false : true);
